requirements.txt
certifi==2021.10.8
Django==2.2.5
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2021.3
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
wincertstore==0.2
